+++ packages/api-client/src/v2/features/app/app-state.ts
@@ -719,7 +719,7 @@ initializeAppEventHandlers({
   onToggleSidebar: () => (isSidebarOpen.value = !isSidebarOpen.value),
 })
 
-export function useAppState(): {
+export type AppState = {
   store: ShallowRef<WorkspaceStore | null>
   sidebar: {
     state: typeof sidebarState
@@ -755,7 +755,12 @@ export function useAppState(): {
   }
   environment: ComputedRef<XScalarEnvironment>
   document: ComputedRef<WorkspaceDocument | null>
-} {
+}
+
+export function useAppState(_router: Router): AppState {
+  if (_router) {
+    router.value = _router
+  }
   return {
     /** Active workspace store */
     store,