+++ packages/api-client/src/v2/features/app/helpers/create-api-client-app.ts
@@ -7,7 +7,7 @@ import { ROUTES } from '@/v2/features/app/helpers/routes'
 import type { ClientPlugin } from '@/v2/helpers/plugins'
 import type { ClientLayout } from '@/v2/types/layout'
 
-export type CreateApiClientOptions = {
+type CreateApiClientOptions = {
   /**
    * The layout of the client, limited to web or desktop in app
    * @see {@link ClientLayout}
@@ -38,14 +38,12 @@ export const createAppRouter = (layout: CreateApiClientOptions['layout']) => {
  * Create the API Client with router and passes in the workspace store as a prop
  */
 export const createApiClientApp = (el: HTMLElement | null, { layout = 'desktop', plugins }: CreateApiClientOptions) => {
-  // Pass in our initial props at the top level
-  const app = createApp(App, { layout, plugins })
-
   // Add the router
   const router = createAppRouter(layout)
-  const state = useAppState()
-  state.router.value = router
+  const state = useAppState(router)
 
+  // Pass in our initial props at the top level
+  const app = createApp(App, { layout, plugins, getAppState: () => state })
   app.use(router)
 
   // Mount the vue app