+++ packages/api-client/src/views/Request/RequestSection/index.ts
@@ -1,3 +1,2 @@
-export { default as RequestSection } from './RequestSection.vue'
-
 export { filterSecurityRequirements } from './helpers/filter-security-requirements'
+export { default as RequestSection } from './RequestSection.vue'