+++ packages/api-client/src/views/Request/RequestSection/types/post-response.ts
@@ -1,7 +1,0 @@
-import type { PostResponseScript, PostResponseScripts } from '@scalar/oas-utils/entities/spec'
-
-export type { PostResponseScript }
-export type PostResponseScriptsProps = {
-  scripts: PostResponseScripts | undefined
-  onUpdate: (scripts: PostResponseScripts | undefined) => void
-}