+++ packages/workspace-store/src/helpers/detect-changes-proxy.ts
@@ -3,16 +3,6 @@ import { isObject } from '@/helpers/general'
 const isDetectChangesProxy = Symbol('isDetectChangesProxy')
 const detectChangesProxyTarget = Symbol('detectChangesProxyTarget')
 
-const isPropsVue = (prop: string | Symbol) => {
-  return (
-    typeof prop === 'symbol' ||
-    prop === '__v_isReactive' ||
-    prop === '__v_isReadonly' ||
-    prop === '__v_raw' ||
-    prop === '__v_skip'
-  )
-}
-
 /**
  * createDetectChangesProxy - Creates a proxy for an object or array that detects and triggers hooks on changes.
  *
@@ -78,7 +68,7 @@ export const createDetectChangesProxy = <T>(
       // Recursively wrap property values in the detect changes proxy
       const value = Reflect.get(target, prop, receiver)
 
-      if (isDetectChangesProxyObject(value) || isPropsVue(prop)) {
+      if (isDetectChangesProxyObject(value)) {
         return value
       }