+++ packages/workspace-store/src/schemas/extensions/security/x-scalar-selected-security.ts
@@ -3,7 +3,7 @@ import { Type } from '@scalar/typebox'
 import { SecurityRequirementObjectRef } from '@/schemas/v3.1/strict/ref-definitions'
 import type { SecurityRequirementObject } from '@/schemas/v3.1/strict/security-requirement'
 
-export const XScalarSelectedSecuritySchema = Type.Object({
+export const XScalarSelectedSecuritySchemaDefinition = Type.Object({
   'x-scalar-selected-security': Type.Optional(
     Type.Object({
       selectedIndex: Type.Integer(),