+++ integrations/fastify/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @scalar/fastify-api-reference
 
+## 1.43.8
+
+## 1.43.7
+
+### Patch Changes
+
+#### Updated Dependencies
+
+- **@scalar/openapi-parser@0.24.1**
+
+- **@scalar/core@0.3.32**
+
+## 1.43.6
+
+### Patch Changes
+
+#### Updated Dependencies
+
+- **@scalar/openapi-parser@0.24.0**
+  - [#7171](https://github.com/scalar/scalar/pull/7171): refactor: new validator class
+
 ## 1.43.5
 
 ## 1.43.4