+++ packages/import/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @scalar/import
 
+## 0.4.45
+
+### Patch Changes
+
+#### Updated Dependencies
+
+- **@scalar/helpers@0.2.8**
+  - [#7751](https://github.com/scalar/scalar/pull/7751): fix: auth persistence
+
 ## 0.4.44
 
 ### Patch Changes