+++ packages/mock-server/docker/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @scalar/mock-server-docker
 
+## 0.1.11
+
+### Patch Changes
+
+#### Updated Dependencies
+
+- **@scalar/mock-server@0.8.15**
+
+- **@scalar/hono-api-reference@0.9.34**
+
+## 0.1.10
+
+### Patch Changes
+
+#### Updated Dependencies
+
+- **@scalar/mock-server@0.8.14**
+
+- **@scalar/hono-api-reference@0.9.34**
+
+## 0.1.9
+
+### Patch Changes
+
+#### Updated Dependencies
+
+- **@scalar/mock-server@0.8.13**
+
+- **@scalar/hono-api-reference@0.9.33**
+
 ## 0.1.8
 
 ### Patch Changes